+++ App/Componentes/funciones.mjs
